New flag to support short node names
This comes from a discussion in hashicorp/docker-consul#69 where it was
observed in some orchestration environments its actually rather clunky
to try and trim the node names in order to make them discoverable via
DNS.

Since Consul warns this is a potential problem, let's provide an easy
way for an operator to address the problem.

The implementation is a dead simple strings.Split and follows the behavior in
[hostname.c][0]. Most of the changes introduced here are around fleshing
out more of the testing in hostname parsing.

[0]: https://sources.debian.org/src/hostname/3.23/hostname.c/#L269